xilpm: versal: client: Fix MISRA 10.3 violation
Fix below MISRA-C violation:
- misra_c_2012_rule_10_3_violation:


Acked-for-series: Jolly Shah <jolly.shah@xilinx.com>

+++ lib/sw_services/xilpm/src/versal/server/xpm_requirement.c
@@ -49,7 +49,7 @@ static XStatus XPmRequirement_Init(XPm_Requirement *Reqm,
 
 	if ((NULL != Params) && (0U != NumParams) && (NumParams <= MAX_REQ_PARAMS)) {
 		(void)XPlmi_MemCpy(Reqm->Params, Params, NumParams * sizeof(*Params));
-		Reqm->NumParams = NumParams;
+		Reqm->NumParams = (u8)NumParams;
 	} else {
 		(void)memset(Reqm->Params, 0, sizeof(Reqm->Params));
 		Reqm->NumParams = 0;
